[CodeGen] Add target hook shouldReMaterializeTrivialRegDef
Adds a target hook `shouldReMaterializeTrivialRegDef` that enables target to specify wether rematerialization of the copy is beneficial. This patch also provide an implementation for AArch64 based on the new subtarget hooks `canLowerToZeroCycleReg[Move|Zeroing]`.

It prepares for a register coalescer optimization to prevent rematerialization of moves where the target supports ZCM.

commit-id:1ce8f604